+++ beacon-chain/rpc/endpoints.go
@@ -200,25 +200,26 @@ func (s *Service) validatorEndpoints(
 	rewardFetcher rewards.BlockRewardsFetcher,
 ) []endpoint {
 	server := &validator.Server{
-		HeadFetcher:            s.cfg.HeadFetcher,
-		TimeFetcher:            s.cfg.GenesisTimeFetcher,
-		SyncChecker:            s.cfg.SyncService,
-		OptimisticModeFetcher:  s.cfg.OptimisticModeFetcher,
-		AttestationCache:       s.cfg.AttestationCache,
-		AttestationsPool:       s.cfg.AttestationsPool,
-		PeerManager:            s.cfg.PeerManager,
-		Broadcaster:            s.cfg.Broadcaster,
-		V1Alpha1Server:         validatorServer,
-		Stater:                 stater,
-		SyncCommitteePool:      s.cfg.SyncCommitteeObjectPool,
-		ChainInfoFetcher:       s.cfg.ChainInfoFetcher,
-		BeaconDB:               s.cfg.BeaconDB,
-		BlockBuilder:           s.cfg.BlockBuilder,
-		OperationNotifier:      s.cfg.OperationNotifier,
-		TrackedValidatorsCache: s.cfg.TrackedValidatorsCache,
-		PayloadIDCache:         s.cfg.PayloadIDCache,
-		CoreService:            coreService,
-		BlockRewardFetcher:     rewardFetcher,
+		HeadFetcher:              s.cfg.HeadFetcher,
+		TimeFetcher:              s.cfg.GenesisTimeFetcher,
+		SyncChecker:              s.cfg.SyncService,
+		OptimisticModeFetcher:    s.cfg.OptimisticModeFetcher,
+		AttestationCache:         s.cfg.AttestationCache,
+		AttestationsPool:         s.cfg.AttestationsPool,
+		PeerManager:              s.cfg.PeerManager,
+		Broadcaster:              s.cfg.Broadcaster,
+		V1Alpha1Server:           validatorServer,
+		Stater:                   stater,
+		SyncCommitteePool:        s.cfg.SyncCommitteeObjectPool,
+		ChainInfoFetcher:         s.cfg.ChainInfoFetcher,
+		BeaconDB:                 s.cfg.BeaconDB,
+		BlockBuilder:             s.cfg.BlockBuilder,
+		OperationNotifier:        s.cfg.OperationNotifier,
+		TrackedValidatorsCache:   s.cfg.TrackedValidatorsCache,
+		PayloadIDCache:           s.cfg.PayloadIDCache,
+		CoreService:              coreService,
+		BlockRewardFetcher:       rewardFetcher,
+		ProposerPreferencesCache: s.cfg.ProposerPreferencesCache,
 	}
 
 	const namespace = "validator"
@@ -399,6 +400,17 @@ func (s *Service) validatorEndpoints(
 				middleware.ContentTypeHandler([]string{api.JsonMediaType}),
 			},
 			handler: server.SyncCommitteeSelections,
+			methods: []string{http.MethodPost},
+		},
+		{
+			template: "/eth/v1/validator/proposer_preferences",
+			name:     namespace + ".SubmitProposerPreferences",
+			middleware: []middleware.Middleware{
+				middleware.ContentTypeHandler([]string{api.JsonMediaType}),
+				middleware.AcceptHeaderHandler([]string{api.JsonMediaType}),
+				middleware.AcceptEncodingHeaderHandler(),
+			},
+			handler: server.SubmitProposerPreferences,
 			methods: []string{http.MethodPost},
 		},
 	}
@@ -759,17 +771,6 @@ func (s *Service) beaconEndpoints(
 			},
 			handler: server.ListProposerPreferences,
 			methods: []string{http.MethodGet},
-		},
-		{
-			template: "/eth/v1/beacon/pool/proposer_preferences",
-			name:     namespace + ".SubmitProposerPreferences",
-			middleware: []middleware.Middleware{
-				middleware.ContentTypeHandler([]string{api.JsonMediaType}),
-				middleware.AcceptHeaderHandler([]string{api.JsonMediaType}),
-				middleware.AcceptEncodingHeaderHandler(),
-			},
-			handler: server.SubmitProposerPreferences,
-			methods: []string{http.MethodPost},
 		},
 		{
 			template: "/eth/v1/beacon/pool/proposer_slashings",
