+++ beacon-chain/rpc/eth/validator/server.go
@@ -19,23 +19,24 @@ import (
 // Server defines a server implementation of the gRPC Validator service,
 // providing RPC endpoints intended for validator clients.
 type Server struct {
-	HeadFetcher            blockchain.HeadFetcher
-	TimeFetcher            blockchain.TimeFetcher
-	SyncChecker            sync.Checker
-	AttestationCache       *cache.AttestationCache
-	AttestationsPool       attestations.Pool
-	PeerManager            p2p.PeerManager
-	Broadcaster            p2p.Broadcaster
-	Stater                 lookup.Stater
-	OptimisticModeFetcher  blockchain.OptimisticModeFetcher
-	SyncCommitteePool      synccommittee.Pool
-	V1Alpha1Server         eth.BeaconNodeValidatorServer
-	ChainInfoFetcher       blockchain.ChainInfoFetcher
-	BeaconDB               db.HeadAccessDatabase
-	BlockBuilder           builder.BlockBuilder
-	OperationNotifier      operation.Notifier
-	CoreService            *core.Service
-	BlockRewardFetcher     rewards.BlockRewardsFetcher
-	TrackedValidatorsCache *cache.TrackedValidatorsCache
-	PayloadIDCache         *cache.PayloadIDCache
+	HeadFetcher              blockchain.HeadFetcher
+	TimeFetcher              blockchain.TimeFetcher
+	SyncChecker              sync.Checker
+	AttestationCache         *cache.AttestationCache
+	AttestationsPool         attestations.Pool
+	PeerManager              p2p.PeerManager
+	Broadcaster              p2p.Broadcaster
+	Stater                   lookup.Stater
+	OptimisticModeFetcher    blockchain.OptimisticModeFetcher
+	SyncCommitteePool        synccommittee.Pool
+	V1Alpha1Server           eth.BeaconNodeValidatorServer
+	ChainInfoFetcher         blockchain.ChainInfoFetcher
+	BeaconDB                 db.HeadAccessDatabase
+	BlockBuilder             builder.BlockBuilder
+	OperationNotifier        operation.Notifier
+	CoreService              *core.Service
+	BlockRewardFetcher       rewards.BlockRewardsFetcher
+	TrackedValidatorsCache   *cache.TrackedValidatorsCache
+	PayloadIDCache           *cache.PayloadIDCache
+	ProposerPreferencesCache *cache.ProposerPreferencesCache
 }
